requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-docx==1.1.0
pydantic==2.5.0
pydantic_core==2.14.1
passlib==1.7.4
bcrypt==4.1.2
python-multipart==0.0.6
requests==2.31.0
